requirements.txt

Flask==1.0.2
python-cas==1.2.0
SQLAlchemy==1.2.9
psycopg2==2.7.5